requirements.txt
django-globus-portal-framework<0.5
